Move mypyc.ops_dict to mypyc.primitives.dict_ops
