+++ mypyc/irbuild/classdef.py
@@ -15,7 +15,7 @@ from mypyc.ops_misc import (
     dataclass_sleight_of_hand, py_setattr_op, pytype_from_template_op, py_calc_meta_op,
     type_object_op, py_hasattr_op, not_implemented_op, true_op
 )
-from mypyc.ops_dict import dict_set_item_op, new_dict_op
+from mypyc.primitives.dict_ops import dict_set_item_op, new_dict_op
 from mypyc.ops_tuple import new_tuple_op
 from mypyc.common import SELF_NAME
 from mypyc.irbuild.util import (
