+++ mypyc/irbuild/function.py
@@ -22,7 +22,7 @@ from mypyc.ops_misc import (
     method_new_op
 )
 from mypyc.ops_exc import raise_exception_with_tb_op
-from mypyc.ops_dict import dict_set_item_op
+from mypyc.primitives.dict_ops import dict_set_item_op
 from mypyc.common import (
     SELF_NAME, ENV_ATTR_NAME, NEXT_LABEL_ATTR_NAME, LAMBDA_NAME, decorator_helper_name
 )
