+++ mypyc/irbuild/ll_builder.py
@@ -33,9 +33,7 @@ from mypyc.ops_list import (
     list_extend_op, list_len_op, new_list_op
 )
 from mypyc.ops_tuple import list_tuple_op, new_tuple_op
-from mypyc.ops_dict import (
-    new_dict_op, dict_update_in_display_op,
-)
+from mypyc.primitives.dict_ops import new_dict_op, dict_update_in_display_op
 from mypyc.ops_misc import (
     none_op, none_object_op, false_op,
     py_getattr_op, py_call_op, py_call_with_kwargs_op, py_method_call_op,
